+++ src/component/layout-container.component.tsx
@@ -16,14 +16,12 @@ import {
   getKeyCombinationsFromActionCodes,
   getLayerShiftPositionCodeMap,
   getModifierKeyPositionCodeMap,
-  nonNullable
+  nonNullable,
 } from "tangent-cc-lib";
 import { HIGHLIGHT_SETTING } from "../const/highlight-setting.const";
 import {
   CC1_DEFAULT_DEVICE_LAYOUT,
-  CC1_LEFT_HAND_ONLY_DEVICE_LAYOUT,
-  CC1_RIGHT_HAND_ONLY_DEVICE_LAYOUT,
-  M4G_DEFAULT_DEVICE_LAYOUT,
+  PRESET_DEVICE_LAYOUTS
 } from "../data/device-layouts";
 import { KEYBOARD_LAYOUTS } from "../data/keyboard-layouts";
 import { useSettingsStore } from "../store/settings-store";
@@ -43,14 +41,9 @@ const LayoutContainerComponent: FC<LayoutContainerProps> = ({ nextText }) => {
   const showThumb3Switch = useSettingsStore.use.showThumb3Switch();
 
   const deviceLayout =
-    [
-      M4G_DEFAULT_DEVICE_LAYOUT,
-      CC1_DEFAULT_DEVICE_LAYOUT,
-      CC1_LEFT_HAND_ONLY_DEVICE_LAYOUT,
-      CC1_RIGHT_HAND_ONLY_DEVICE_LAYOUT,
-      ...customDeviceLayouts,
-    ].find((deviceLayout) => deviceLayout.id === layout) ||
-    CC1_DEFAULT_DEVICE_LAYOUT;
+    [...PRESET_DEVICE_LAYOUTS, ...customDeviceLayouts].find(
+      (deviceLayout) => deviceLayout.id === layout,
+    ) || CC1_DEFAULT_DEVICE_LAYOUT;
   const selectedKeyboardLayout = KEYBOARD_LAYOUTS.find(
     (k) => k.id === selectedKeyboardLayoutId,
   ) as KeyboardLayout;